Rename files for DynamicCoin
