+++ src/components/uielements/coins/dynamicCoin/DynamicCoin.style.tsx
